transition to the library SparkFun for sensors BME280 and CCS811

+++ include/json.h
@@ -4,4 +4,4 @@ String jsonRead(String &, String);
 // int jsonRead(String &, String);
 
 String jsonWrite(String &, String, String);
-String jsonWrite(String &, String, int);
+String jsonWrite(String &, String, float);
